backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
asyncpg==0.30.0
pydantic-settings==2.8.1
python-multipart==0.0.20
celery==5.4.0
redis==5.2.1
reportlab==4.3.1
pytest==8.3.5
httpx==0.28.1
boto3==1.37.30
azure-identity==1.21.0
azure-mgmt-subscription==3.1.1
azure-mgmt-storage==21.2.1
google-auth==2.38.0
google-cloud-storage==3.1.0
google-cloud-resource-manager==1.14.2
